Fixed DataTypes i think, or i broke them more. stay tuned

+++ piggybank/src/piggybank/ChangeActions.java
@@ -3,8 +3,8 @@ package piggybank;
 public abstract class ChangeActions extends MoneyActions{
 
 	private String type;
-	public abstract double getAmmount();
-	public abstract int getValue();
+	public abstract int getAmmount();
+	public abstract double getValue();
 
 	public ChangeActions(){
 		this.type = "coin.";
